% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_heart.R
\name{heart_assert_required_cols}
\alias{heart_assert_required_cols}
\title{Validate that required columns exist}
\usage{
heart_assert_required_cols(df, required)
}
\arguments{
\item{df}{data.frame.}

\item{required}{Required column names.}
}
\value{
Invisible TRUE or error.
}
\description{
Validate that required columns exist
}
\examples{
b <- heart_load_bundle()
newdata <- heart_example_data(include_outcome = FALSE, n = 1)
heart_assert_required_cols(newdata, b$required_cols)
}
