% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_heart.R
\name{heart_calibration}
\alias{heart_calibration}
\title{Calibration bins and Expected Calibration Error (ECE)}
\usage{
heart_calibration(y_true, prob_pos, positive_level = "Presence", bins = 10)
}
\arguments{
\item{y_true}{True labels.}

\item{prob_pos}{Positive probabilities.}

\item{positive_level}{Positive label.}

\item{bins}{Number of bins.}
}
\value{
list(cal_df, ece)
}
\description{
Calibration bins and Expected Calibration Error (ECE)
}
\examples{
b <- heart_load_bundle()
dat <- heart_example_data(include_outcome = TRUE, n = 80)
y <- factor(ifelse(dat$HeartDisease == 1, b$positive_level, b$negative_level),
            levels = c(b$positive_level, b$negative_level))
x <- dat; x$HeartDisease <- NULL
p <- heart_predict_proba(x, b, model = "topk")
out <- heart_calibration(y, p, positive_level = b$positive_level, bins = 10)
out$ece
head(out$cal_df)
}
