% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:model_evaluation_heart.R
\name{heart_eval_threshold}
\alias{heart_eval_threshold}
\title{Compute classification metrics at a fixed threshold}
\usage{
heart_eval_threshold(y_true, prob_pos, threshold, positive_level = "Presence")
}
\arguments{
\item{y_true}{Factor/character of true labels.}

\item{prob_pos}{Numeric vector of positive probabilities.}

\item{threshold}{Threshold for positive.}

\item{positive_level}{Positive class label.}
}
\value{
A named list of metrics.
}
\description{
Compute classification metrics at a fixed threshold
}
