% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_heart.R
\name{heart_glmnet_coef_importance}
\alias{heart_glmnet_coef_importance}
\title{Extract glmnet coefficient importance (absolute coefficients at best lambda)}
\usage{
heart_glmnet_coef_importance(glm_caret)
}
\arguments{
\item{glm_caret}{A caret::train object fitted with method='glmnet'.}
}
\value{
tibble(Feature, Coef, Importance)
}
\description{
Extract glmnet coefficient importance (absolute coefficients at best lambda)
}
\keyword{internal}
