% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:predict_heart.R
\name{heart_load_bundle}
\alias{heart_load_bundle}
\title{Load the pre-trained model bundle shipped with the package}
\usage{
heart_load_bundle(path = NULL)
}
\arguments{
\item{path}{Optional path to a bundle .rds. If NULL, loads from package extdata.}
}
\value{
A list (bundle) containing trained models + preprocessing objects.
}
\description{
Load the pre-trained model bundle shipped with the package
}
