% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:model_interpretation_heart.R
\name{heart_plot_importance}
\alias{heart_plot_importance}
\title{Plot feature importance as a horizontal bar chart}
\usage{
heart_plot_importance(imp_tbl, title, ylab = "Importance")
}
\arguments{
\item{imp_tbl}{A tibble with Variable and Importance.}

\item{title}{Plot title.}

\item{ylab}{Y-axis label.}
}
\value{
A ggplot object.
}
\description{
Plot feature importance as a horizontal bar chart
}
