% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{heart_required_cols_default}
\alias{heart_required_cols_default}
\title{Get prediction schema (required columns)}
\usage{
heart_required_cols_default()
}
\value{
Character vector of required predictor columns.
}
\description{
Get prediction schema (required columns)
}
\keyword{internal}
