% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:model_interpretation_heart.R
\name{heart_theme}
\alias{heart_theme}
\title{Package plotting theme}
\usage{
heart_theme(base_size = 14)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot2 theme.
}
\description{
Package plotting theme
}
