% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartPredictionR-package.R
\docType{package}
\name{heartPredictionR-package}
\alias{heartPredictionR}
\alias{heartPredictionR-package}
\title{heartPredictionR: Heart Disease Risk Prediction with Machine Learning}
\description{
\code{heartPredictionR} provides a reproducible interface for predicting heart disease risk
from routine clinical features using a pre-trained model bundle shipped with the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tian233-bot/heartPredictR}
  \item Report bugs at \url{https://github.com/tian233-bot/heartPredictR/issues}
}

}
\author{
\strong{Maintainer}: Haotian Guo \email{Haotian.Guo23@student.xjtlu.edu.cn}

}
\keyword{internal}
