% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:model_evaluation_heart.R
\name{heart_calibration}
\alias{heart_calibration}
\title{Calibration bins and expected calibration error (ECE)}
\usage{
heart_calibration(y_true, prob_pos, positive_level = "Presence", bins = 10)
}
\arguments{
\item{y_true}{true labels.}

\item{prob_pos}{positive probabilities.}

\item{positive_level}{positive label.}

\item{bins}{number of bins.}
}
\value{
list(cal_df, ece).
}
\description{
Calibration bins and expected calibration error (ECE)
}
