% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_heart.R
\name{heart_coerce_inputs}
\alias{heart_coerce_inputs}
\title{Coerce raw inputs to model-friendly types}
\usage{
heart_coerce_inputs(
  df,
  numeric_cols = c("Age", "RestingBP", "Cholesterol", "MaxHR", "Oldpeak"),
  cat_cols = c("Sex", "ChestPainType", "FastingBS", "RestingECG", "ExerciseAngina",
    "ST_Slope")
)
}
\arguments{
\item{df}{data.frame of predictors.}

\item{numeric_cols}{Numeric feature names.}

\item{cat_cols}{Categorical feature names.}
}
\value{
A tibble with coerced columns.
}
\description{
Numeric columns are coerced to numeric (hard fail if NA is introduced).
Categorical columns are coerced to character. (The model/recipe will handle factorisation.)
}
