% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_heart.R
\name{heart_importance_logistic}
\alias{heart_importance_logistic}
\title{Aggregate one-hot dummy importances back to raw variables (logistic model)}
\usage{
heart_importance_logistic(
  glm_caret,
  top_n = 20,
  cat_prefix = c("Sex", "ChestPainType", "FastingBS", "RestingECG", "ExerciseAngina",
    "ST_Slope"),
  how = c("sum", "max")
)
}
\arguments{
\item{glm_caret}{A caret::train(glmnet) object.}

\item{top_n}{Top N raw variables.}

\item{cat_prefix}{Categorical prefixes to aggregate (e.g. Sex_*).}

\item{how}{Aggregation method: sum or max.}
}
\value{
tibble(Variable, Importance)
}
\description{
Aggregate one-hot dummy importances back to raw variables (logistic model)
}
