% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:model_interpretation_heart.R
\name{heart_importance_rf}
\alias{heart_importance_rf}
\title{Extract RF permutation importance (vip::vi)}
\usage{
heart_importance_rf(ranger_model, top_n = 20)
}
\arguments{
\item{ranger_model}{A ranger model object (e.g., caret train finalModel).}

\item{top_n}{Top N features.}
}
\value{
A tibble with Variable and Importance.
}
\description{
Extract RF permutation importance (vip::vi)
}
