% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_heart.R
\name{heart_importance_rf}
\alias{heart_importance_rf}
\title{Extract Random Forest permutation importance (vip::vi)}
\usage{
heart_importance_rf(ranger_model, top_n = 20)
}
\arguments{
\item{ranger_model}{A fitted ranger model.}

\item{top_n}{Top N variables.}
}
\value{
tibble(Variable, Importance)
}
\description{
Extract Random Forest permutation importance (vip::vi)
}
