% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_heart.R
\name{heart_load_bundle}
\alias{heart_load_bundle}
\title{Load the pre-trained model bundle shipped with the package}
\usage{
heart_load_bundle(
  path = NULL,
  filename = "heart_models_RECIPES_STRICT_bundle.rds"
)
}
\arguments{
\item{path}{Optional path to a bundle \code{.rds}. If NULL, load from package extdata.}

\item{filename}{Bundle filename inside package extdata (advanced).}
}
\value{
A list (bundle) containing trained models and metadata.
}
\description{
By default the bundle is loaded from \code{extdata} in the installed package.
You can also provide a custom path to an \code{.rds} bundle file.
}
