% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_heart.R
\name{heart_plot_importance}
\alias{heart_plot_importance}
\title{Plot importance bar chart}
\usage{
heart_plot_importance(imp_tbl, title, ylab = "Importance")
}
\arguments{
\item{imp_tbl}{tibble with Variable, Importance.}

\item{title}{Plot title.}

\item{ylab}{y-axis label.}
}
\value{
ggplot object.
}
\description{
Plot importance bar chart
}
