% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_heart.R
\name{heart_theme}
\alias{heart_theme}
\title{Package ggplot theme}
\usage{
heart_theme(base_size = 14)
}
\arguments{
\item{base_size}{base font size.}
}
\value{
ggplot theme.
}
\description{
Package ggplot theme
}
\examples{
heart_theme()
}
